cfclient_cache (cfclient v2.0.1)

Functions to manage cache of Flag and Segment data from server.

Types

config/0

-type config() :: cfclient:config().

flag/0

-type flag() :: cfclient_evaluator:flag().

segment/0

-type segment() :: cfapi_segment:cfapi_segment().

Functions

cache_flag(Value)

-spec cache_flag(flag()) -> ok | {error, outdated}.

cache_flag(Value, Config)

-spec cache_flag(flag(), config()) -> ok | {error, outdated}.

cache_segment(Value)

-spec cache_segment(segment()) -> ok | {error, outdated}.

cache_segment(Value, Config)

-spec cache_segment(segment(), config()) -> ok | {error, outdated}.

get_value(_)

-spec get_value({flag, binary()} | {segment, binary()}) -> {ok, flag() | segment()} | {error, undefined}.

Get Flag or Segment from cache.

get_value(_, Config)

-spec get_value({flag, binary()} | {segment, binary()}, config()) ->
             {ok, flag() | segment()} | {error, undefined}.

set_pid(_)

-spec set_pid(pid()) -> ok.

set_value(_, Value)

-spec set_value({flag, binary()} | {segment, binary()}, flag() | segment()) -> ok | {error, outdated}.

Store flag or segment into cache with new value.