cfclient_instance (cfclient v1.0.0)

Feature flags client instance.

It creates the ETS tables used to cache flag data from the server and flag usage metrics. It runs periodic tasks to pull data from the server and send metrics to it.
An default instance is started by the cfclient application. Additional instances can be started if multiple Harness projects need to be used. project.

Functions

handle_call(_, From, State)

handle_cast(_, State)

handle_info(_, Config)

init(Args)

start_link(Args)

-spec start_link(proplists:proplist()) -> {ok, pid()} | ignore | {error, term()}.

stop(Config)

-spec stop(map()) -> ok | {error, not_found, term()}.